moved handover to extension module
