+++ beacon_api/extensions/__init__.py
@@ -1,0 +1,1 @@
+"""Extension Module."""
